Update patch set 1
Patch Set 1: Presubmit-Ready+1

This is the filemap tool into which I had incorporated all prior feedback.
Sandeep has asked me to put it in AOSP instead of the internal Android master.

I examined the libfiemap_writer and decided that it was simplest just to check-in this tool as is instead of trying to adapt it to that library. Maybe it can adapted at a later time, but there is little value in doing so. The tool as it is can be easily used standalone on non-Android Linux systems without much effort.

Patch-set: 1
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1428004 <1428004@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1