Fix: bionic files are unabled to be pushed to the device
This CL fixes a problem that `adb push` and `adb sync` do not work for
the bionic files that were moved to the runtime APEX.

With [1], the bionic libraries and the dynanic linker are moved to the
runtime APEX. However, in order to provide the same path, e.g.,
/system/lib/libc.so, to all processes, the files in the runtime APEX are
bind-mounted to the old paths. This is causing problem when pushing
files to the paths which are now mountpoints and not regular files.

To solve the problem, when adbd executes the file sync service, it
temporarily creates a new mount namespace and unmounts the bionic files
to make the paths back to regular files. Since the mount namespace is
created only for the thread executing the file sync service, the
unmounts are not done for other processes and other threads in adbd.

[1] https://android-review.googlesource.com/q/topic:move_bionic_to_apex
Bug: 122737045
Test: adb root; adb remount; adb sync

Change-Id: I628f510c62df68709b7b57e54e44671b9a02a49e