Add libc_headers to header_libs for libpropertyinfoparser
This commit adds `libc_headers` to `header_libs` for
`libpropertyinfoparser`.  Before this commit, `libpropertyinfoparser`
relies on `libc` that are silently added to `system_shared_libs`.  This
will no longer work if the build system always respects the
`system_shared_libs` that are specified in `Android.bp`.  Thus, we
should depend on `libc_headers` instead.

Bug: 119086738
Test: make checkbuild
Change-Id: I2d416c6db7f42f12ad20d98f4048753254f5e0b6