Update patch set 2
Patch Set 2:

(2 comments)

I'm okay either way too.  I know we had bugs filed in the past about the kernel panic, so maybe better to not have it?  Though if the timeout is 30 seconds, then a kernel panic, and presumably a ramdump that would be uploaded would be a feature not a bug, since we really shouldn't ever hit this time out.

The real question; is should this be expanded to cover the reboot case too or continue with just the fsck case?  I'm guessing we do want this to be expanded to everything but thermal shutdown, right?

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>