Fix: bionic files are unabled to be pushed to the device
This CL fixes a problem that `adb push` and `adb sync` do not work for
the bionic files that were moved to the runtime APEX.

With [1], the bionic libraries and the dynanic linker are moved to the
runtime APEX. However, in order to provide the same path, e.g.,
/system/lib/libc.so, to all processes, the files in the runtime APEX are
bind-mounted to the old paths. This is causing problem when pushing
files to the paths which are now mountpoints and not regular files.

The solution to this problem are two folds:

1) Init no longer bind mounts the bootstrap bionic to the mountpoints.
Instead, it makes following three paths which are the parent directories
of the bionic files as private mount points
- system/lib
- system/lib64
- system/bin

Bionic from the runtime APEX are bind-mounted to the ordinary paths
(/system/lib/libc.so, etc.) as before.

2) when adbd executes the file sync service service, it temporarily
creates a new mount namespace and unmounts the new private mount points
to reveal the underlying raw filesystem. The sync is done on top of the
raw filesystem.

In addition, since we are introducing mount points inside the system
partition, they are also explicitly remounted during `adb remount`.

[1] https://android-review.googlesource.com/q/topic:move_bionic_to_apex
Bug: 122737045
Test: adb root; adb remount; adb sync

Change-Id: I392bd0838854f27dfc93893c9c7fdd777316c1c0