Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review+1
> 
> Merging this to unblock development SGTM, as long as the plan is to revert this with a proper fix.
> 
> (For the record, my complaints are that it's bad to have to keep a list of directories in sync across 3 files, and that `adb pull /system/lib/libc.so` should give you the same output as `adb shell cat /system/lib/libc.so > libc.so`)

Thank you.

BTW, this fix might not be needed as the problematic change will be reverted due to some 3p app crashing issue.
https://android-review.googlesource.com/q/topic:%22revert_move_bionic_to_apex%22+(status:open%20OR%20status:merged)

The inconsistent content problem that you just mentioned is actually due to the kernel bug in sdcardfs (b/122428178). Because of the bug, we are currently *not* using multiple namespaces across post and pre-apexd processes. As a result, the libc from runtime APEX which is not for adbd and its children are visible to adbd. This will be fixed when we correctly have multiple namespaces. But given the tight time constraint, I am very skeptical that we can make it on time for Q.

Patch-set: 9
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0