Update patch set 7
Patch Set 7:

(22 comments)

A bunch of style comments + a few design ones.  I think the biggest question that I have is can we move the code that actually mounts and sets up the cgroups into this file, and configure them through task_profiles.json?  There's no reason that they "have" to be done in init scripts, and if we do it all via task_profiles.json, we can simplify the code significantly, since we both would know exactly when all of the cgroups are mounted and exactly what paths and mount points were used to set them up.

Just an idea, but I'm thinking it might save a lot of complexity.

Patch-set: 7
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>