Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

As far as I know there is no reason for this change. SocketListener was converted to poll() a while ago, so should be no calls to select() any more. Thus it is safe to use fd numbers that are > 1024.

In fact, this change will cause a regression: if there are more than 1024 fds open, then with current code, new requests will succeed. But if this change is merged, if there are more than 1024 fds open, then new requests will fail.So I don't think we should take this change as is.

I'm not in the OWNERS file and cannot approve a change to this code, so I defer to Elliott. But if you want something like this change to be merged so you can backport it to P devices, you should make the limit be a separate configurable #define. Alternatively, for DnsProxyListener you could lower the OperationLimiter thresholds, or add another OperationLimiter for total queries.

Patch-set: 3
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1