fs_mgr: overlayfs: erofs is in staging
erofs is in upstream staging, and we see some Android partners
considering it as an option instead of squashfs.

Let's proactively deal with them for overlayfs handling to support
adb remount during development.  This does _not_ mean we support the
decision by our partners to use erofs, as we have not evaluated the
performance.

Test: compile
Change-Id: I80ea3a33b9da6684ee95489132bc8a88b2310ac8