fs_mgr: remount add -R flag
The -R flag tells remount it can reboot to disable verity or to
run fsck on an ext4 deduped filesystem, or both.

Testing may include a manual component because adb-remount-test.sh
can not always run adb enable-verity since it can not be guaranteed
to reboot after such an action unless we are sure the device uses
overlayfs.  The test condition requires that the device has verity
enabled before running, so we try to put the device into that state
after the completion of the test.

Test: adb-remount-test.sh
Bug: 122602260
Change-Id: I6ce4372532d9b933dcca9e2bec544d525b76c4d9