fs_mgr: remount add -R flag
The -R flag tells remount it can reboot to disable verity or to
run fsck on an ext4 deduped filesystem, or both.

Since these specific operations should not be performed on-device
except in a debug scenario, we will limit the action to process
parentage pointing eventually to adbd which means done over the
physical link.  This will prevent random on-device applications
from being able to perform these operations.

Testing may include a manual component because adb-remount-test.sh
can not always run adb enable-verity since it can not be guaranteed
to reboot after such an action unless we are sure the device uses
overlayfs.  The test condition requires that the device has verity
enabled before running, so we try to put the device into that state
after the completion of the test.

Only recognizes chained avb.

Test: adb-remount-test.sh
Bug: 122602260
Change-Id: I6ce4372532d9b933dcca9e2bec544d525b76c4d9