Update patch set 2
Patch Set 2:

This seems like it limits applications to exactly one retry strategy.

It might be valuable for there to be a connect_with_retry function or similar wrapper function, but hardcoding all usage of tipc_connect to auto-retry seems unnecessarily restrictive.

Caveat: I'm not super familiar with this part of the code.

Patch-set: 2
CC: Gerrit User 1416371 <1416371@85c56323-6fa9-3386-8a01-6480fb634889>