Update patch set 1
Patch Set 1: Presubmit-Ready+1

(1 comment)

ext4_utils is still needed at least for the static init due to fs_mgr; mkbootimg is already gone.

I saw the docs that this was to be used in tests, but it seems like that either never happened or was removed, since I don't see any references.

Patch-set: 1
Label: Presubmit-Ready=+1