fs_mgr_remount: support legacy devices (marlin)
Deal with first version of verity for legacy products.  Expand the
tests to deal with the boot environment for marlin.  Recognize that
older overlayfs drivers do not report to /sys/module and the
parsing /proc/filesystem is another place to interrogate this.

Adds support for the legacy avb implementation.

NB: Running test to completion is difficult because marlin's USB
    driver is flakey enough through the multitude of reboots and
    may not reconnect.  The tester will have to notice when a reboot
    is stalling and manually disconnect and reconnect the USB
    connection to trigger discovery and to continue through the
    test sequences.  To make this easier, report when we are
    waiting for the device to make it easier to babysit.

Test: system/core/fs_mgr/tests/adb-remount-test.sh
Bug: 120448575
Bug: 123079041
Change-Id: I5fc5f01b4e4788ac57541cb5235f7ac4e4284d71