Update patch set 6
Patch Set 6:

> Patch Set 6:
> 
> > Patch Set 6:
> > 
> > You probably want to move this someplace earlier than post-fs-data.  A lot of processes will have started before then.
> > 
> 
> I can move it to early-init then
> 
> > What is long term plan for this?  I assume once this is always set to 'true', then it will be dropped from this file entirely and in libcutils, the default will be 'true' unless a developer overrides it via adb?
> 
> Yes that's right

Actually, this property will never be set to true during init time, if it is set to true - then it will be removed as you pointed.

It is only set to true for development purposes (set to true followed by shell stop/start). So I believe it is fine if we leave it in this location (and don't need to move it to earlier)

Patch-set: 6
Reviewer: Gerrit User 1124848 <1124848@85c56323-6fa9-3386-8a01-6480fb634889>