Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > No std::string?
> > 
> > Looks like we did add libbase as a dependency in https://android-review.googlesource.com/c/platform/system/core/+/855149/ so StartsWith() is back on the table too.
> 
> Braver souls :-) so if this happened with no build errors on all platforms (including libcutils for vendor?), sounds like we need to convert libcutils to C++!

Indeed.  It's been ready for C++ for a long while; libbase is the only new thing.  I think in general we've been trying to move people away from libcutils in general though, so there isn't a huge reward in converting the lesser used parts.  But since this code is used and is being modified, it'll pay to convert it.

Patch-set: 7
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>