Update patch set 11
Patch Set 11: Code-Review+1

(2 comments)

Especially now if we're remounting all; I think we should get `adb remount` to use this.  Should be as simple as removing that code and replacing it with a call out to this process.  I chatted with jmgao@ about this and he's on board too.

Patch-set: 11
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1