liblog: Set the close-on-exec (FD_CLOEXEC) flag on the created socket file descriptor
Socket(file descriptor) created in liblog will be unintentionally inherited(or so-called leaked) to child process(created by fork/vfork/system/popen etc). And then, even after the parent process has closed that socket, it will not be actually "closed" due to the child process also has a file descriptor referring to it.  This will make the other end point of socket be ignorant of the truth this end point being stopped, and hence the other end point may be always blocked due to EOF or pipe error (SIGPIPE or EPIPE) is impossible.

Test: monkey test
Change-Id: Idfa6d3b0181488bc1db97456fcd24d198083ae7d