Update patch set 8
Patch Set 8:

(TH failure is a breakage at TOT)

I didn't think that fs_mgr_overlayfs_verity_enabled_list() played a role in the duplicate_or_more_specific logic, especially since it only returns the partitions with verity enabled.  None of the current code paths directly play that role.

In general, I don't think that's the right function to rely on for that logic, since it, in turn just reads the default fstab for its mount information.  If the intention is to read the default fstab to check duplicate/more logical partitions, I think we should take this CL as is and then explicitly read that fstab for the duplicate/more specific checks. 
 Otherwise, maybe it'd be better to read /proc/mounts, since that will have the actual status of what mounts we currently have.

Patch-set: 8
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>