Update patch set 1
Patch Set 1:

I see that this is pretty trivially OK in existing code, but if the primary customer of this is users of unique_lock, it might be a good idea to assert `unique_lock.owns_lock()` in the ctor/dtor here if we have access to e.g. DCHECK.

If we can't take a unique_lock, we can also try `DCHECK(!mtx.try_lock())`. Probably a bit higher overhead, but if it's under a DCHECK anyway...

Nothing that should block the submission of this as-is. Just ideas for maybe the future.

Patch-set: 1
Reviewer: Gerrit User 1096946 <1096946@85c56323-6fa9-3386-8a01-6480fb634889>