Update patch set 1
Patch Set 1:

_why_ do we even have to add anything? can't you just have the build
system automatically implement your advice of "When in doubt, put
gtests in device-tests and APK tests in general-tests", and only
require boilerplate if that's _not_ the case? (which would be when?
the docs don't make it clear. do you actually mean "anything with
native code should be device-tests, and anything Java-only should be
general-tests"? so cc_test should automatically be in device-tests,
and we should do this in the build system?)

Patch-set: 1
Tag: mailMessageId=<CAJgzZooWdLLLGBXDJCytSa6Fvvgt7pteEWKcpLQXSrfYJt67-Q@mail.gmail.com>