Update patch set 3
Patch Set 3: Autosubmit+1

> Patch Set 3: Code-Review+2
> 
> Just wondering if we should check have this code only be on userdebug/eng builds.  We won't want to ever allow this for user builds, right?

Thanks for clarify this.
Currently we don't check build variants, we check whether the device is unlocked or not.

e.g., 'adb disable-verity' is only allowed:
  1) When device is unlocked and
  2) 'adb root' is allowed (which isn't allowed in USER build though). 

https://android.googlesource.com/platform/system/core/+/refs/heads/master/adb/daemon/set_verity_enable_state_service.cpp#143

When the disabled bit is set, it will have 'verification error', which is also not allowed when the device is locked.
https://android.googlesource.com/platform/system/core/+/refs/heads/master/fs_mgr/libfs_avb/fs_avb.cpp#287

As there is no security guarantee when the device is unlocked, it should be fine.

Patch-set: 3
Reviewer: Gerrit User 1080813 <1080813@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Autosubmit=+1
Label: Presubmit-Ready=0