Update patch set 1
Patch Set 1:

having the configuration in the same place as other configuration is
definitely better than inventing something new.

but

* having *convention* rather than configuration

* having "just works" defaults for most users

are even better.

should be a warning at least though.

this seems like the kind of backwards thinking we used to get from the
old test team 10 years ago, asking us for fewer tests so test runs
would be quicker. why even write tests if we're not running them as
soon and as often as possible? i'd suggest instead that we should
*remove* any tests that we're not running all the time. seriously. a
test that isn't being run is actively harmful, because it gives us a
false sense of security. i've only just learned from this thread that
changes to external/arm-optimized-routines aren't running the libm
tests thanks to TEST_MAPPING. so for months now we've had this false
sense of security that things are better tested than they are.

at a minimum we need to really loudly complain if something's in
TEST_MAPPING but didn't actually run. (and this needs to be visible in
treehugger, because although, yes, we can get into the habit of
running atest before we check in, the truth is that stuff breaks over
time... anyone can edit a .bp file *in a different project* and
silently break your TEST_MAPPING right now :-( )

(don't get me wrong: TEST_MAPPING is one of the better things to have
come out of the new test team, second only to the flake monitoring,
but i do think it's in that "dangerous" valley between "we know it's
useless so we're super careful" [like all the old stuff] and "we rely
on it, and it's safe to do so because it's always up-front about
what's going on".)

Patch-set: 1
Tag: mailMessageId=<CAJgzZootAYWrY0EZED5kvBVL1qixfYMy5P5L7e5AjK9L6Ye1CQ@mail.gmail.com>