Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> having the configuration in the same place as other configuration is
> definitely better than inventing something new.
> 
> but
> 
> * having *convention* rather than configuration
> 
> * having "just works" defaults for most users
> 
> are even better.
> 
> should be a warning at least though.
> 
> this seems like the kind of backwards thinking we used to get from the
> old test team 10 years ago, asking us for fewer tests so test runs
> would be quicker. why even write tests if we're not running them as
> soon and as often as possible? i'd suggest instead that we should
> *remove* any tests that we're not running all the time. seriously. a
> test that isn't being run is actively harmful, because it gives us a
> false sense of security. i've only just learned from this thread that
> changes to external/arm-optimized-routines aren't running the libm
> tests thanks to TEST_MAPPING. so for months now we've had this false
> sense of security that things are better tested than they are.
> 
> at a minimum we need to really loudly complain if something's in
> TEST_MAPPING but didn't actually run. (and this needs to be visible in
> treehugger, because although, yes, we can get into the habit of
> running atest before we check in, the truth is that stuff breaks over
> time... anyone can edit a .bp file *in a different project* and
> silently break your TEST_MAPPING right now :-( )
> 
> (don't get me wrong: TEST_MAPPING is one of the better things to have
> come out of the new test team, second only to the flake monitoring,
> but i do think it's in that "dangerous" valley between "we know it's
> useless so we're super careful" [like all the old stuff] and "we rely
> on it, and it's safe to do so because it's always up-front about
> what's going on".)

Thanks for this user feedback and it's good timing as I'm in the process of shuffling some more resources into Test Mapping as it's really just been Dan Shi who built all of Test Mapping so far. Our team's mission has been to drive tests and tools to "just work" as much as possible and that's what we've been trying to push for.

The false sense of security was not something we thought of, so I'll arrange for us to do an audit of how bad this issue is, get it fixed and get some sort of mechanism up to warn users with no extra steps. b/124446646

Feel free to reach out if there's more situations we did not think out end-to-end so we adjust accordingly.

Patch-set: 1
Reviewer: Gerrit User 1083173 <1083173@85c56323-6fa9-3386-8a01-6480fb634889>