Update patch set 3
Patch Set 3:

But you're still using it on host even with that issue?  I guess if it runs into that issue it'll just fail the test?

(What I'm getting at here, is maybe better to not use it at all, if possible, until we get the right version of libprotobuf updated?)

Patch-set: 3
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>