Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> That's what I figure too; I just worry about including code that we know is broken at least until we update protobuf.  Probably fine though, especially if we know if we're getting that update soon (I hope?)

Yeah; I don't want to depend on the process of protobuf being updated or not, because this needs to go before Q. I wrote the test so that it should (?) run with older and newer protobuf nicely. It's just that with a newer protobuf:
- The test can be written more elegantly
- The parser can share some common code.

> But in any case, let's get this moved to system/extras or someplace else that's not directly in libprocessgroups.

Sure. I'll upload patches shortly.

Patch-set: 3
Reviewer: Gerrit User 1121838 <1121838@85c56323-6fa9-3386-8a01-6480fb634889>