Update patch set 2
Patch Set 2:

(1 comment)

> > the ability to run a shell/python script from TEST_MAPPING sounds useful, but should probably be coordinated with the engprod folks so we have one way to do this, usable by everyone.
> 
> python host testing already exists. See symbol lib testing.
> 
> (I don't think shell does)

Is the policy that it must be written in python (or Java)?

The _lack_ of shell support is probably on purpose (I think I am the _only_ person that writes shell scripts, it may be considered a dying breed). But C++ is also not unified (gTest looks like an outlier).

Patch-set: 2
CC: Gerrit User 1014491 <1014491@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0