Update patch set 12
Patch Set 12:

If the goal is to guarantee that the parent process tree includes adbd in a non-robust yet simple way, perhaps environment variables are an easier way to do this. Have adbd set an environment variable ANDROID_ADBD_ENV and have your code check for the existence of the environment variable. 

This has two advantages:

1) Less intrusive and simpler to implement
2) You can setup source code alerts to notify you whenever anyone checks in that magic "ANDROID_ADBD_ENV" string.

And it has the same pseudo-security properties as the current implementation.

There's a lot of code and complexity here, and the complexity will make it harder to iterate on this code. I'd prefer that we not check this code in entirely, but if we can't avoid that, then I want to make the implementation as simple and self-contained as possible.

Patch-set: 12
CC: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>