Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > still seems to be failing, even after your change?
> > 
> > https://sponge.corp.google.com/target?id=129e77fd-39ec-486c-8005-7fb97fde209b&target=x86+libutils_test&searchFor=&show=FAILED&sortBy=STATUS
> > 
> > do i need to s/required:/data:/ in the Android.bp?
> 
> I am not familiar with the build rules required to get the libs packaged with. Dan, Simran?

yes, in system/core/TEST_MAPPING, the test is added as a device test. So any shared library needs to be either pre-installed on device or be part of test data that can be copied over to the device during the test.

Patch-set: 1
Reviewer: Gerrit User 1088076 <1088076@85c56323-6fa9-3386-8a01-6480fb634889>