Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Patch Set 4:
> > 
> > > Patch Set 4:
> > > 
> > > > Patch Set 4:
> > > > 
> > > > > It seems to me that the right way to do this is to specify the port range 
> > > > >  with ADB_EMULATOR_PORT_MIN/MAX, not the number of transports with an arbitrary hard coded 
> > > > 
> > > > If the number of allocated ports ever changes, setting the assumption that there are 2 ports per emulator in the port min/max will mess things up. It seems that ADB_EMULATOR_PORT_BASE and ADB_EMULATOR_MAX would be more future-proof. Perhaps - if we're messing with 5554 - an ADB_EMULATOR_PORT_MULTIPLIER would completely express the algorithm.
> > > 
> > > I think the long term plan is to make the emulator register with adb instead of having adb blindly attempt to connect to a range of ports (+lfy to correct me if I'm wrong). Until then, I think it's unlikely for us to add more ports (and if we do, we could just add them starting at a different base).
> > 
> > Yeah thats right + Agreed with the approach in this CL
> 
> Do you prefer this approach to _MIN and _MAX env vars? That seems cleaner to me, but I don't have a super strong opinion.

Not a super strong opinion here either, but:

I have some concerns about specifying a range, because stemming from the original hard coded port, that's (Hyrum's Law) resulted in an ecosystem of things depending on that such as emulator. It's not really guaranteed that everyone who depends on this will update in time, and it's not guaranteed, now that the base is not hardcoded, that it will change/fork in various ways from the device side.

Patch-set: 4
Reviewer: Gerrit User 1082465 <1082465@85c56323-6fa9-3386-8a01-6480fb634889>