Update patch set 4
Patch Set 4:

> Our servers can only handle half the number of emulators as there are entries in /proc/cpuinfo. Although our scripts should decide how many emulators to ask for, having a max on the total number of emulators in case of broken logic is precisely what we'd want.

It seems like a mistake to try to enforce this limit in adb, though. This doesn't actually limit the number of emulators you can start, it just makes them not show up in adb. When the emulator starts telling adb what port it's started on, I expect that we're going to ignore the limit completely for those emulators.

Patch-set: 4
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>