Internalize subclasses of Memory
There are many subclasses of the Memory class and the overwhelming
majority of them don't need to be exposed externally. We move all of
them to internal headers. This dramatically reduces the exposed API
surface and makes the code more modular.

Also, remove the Offline code from libbacktrace. It's not used anywhere.

Test: Unit tests pass, clean tree still builds
Change-Id: I55dacdf080daba0bfe65c1ad53a4b326bb482e83