adb: eliminate reboot message for overlayfs-only adjustment
There is a subtle difference between the state where overlayfs is
mounted late as an effective afterthought during disable-verity at
runtime and when properly applied in first stage init.  In the former
the development is still usable except for a subset of core native
components and thus the developer may take the shortcut of not
rebooting.  Instead of a mandatory reboot request ...

Developers are welcome to punish themselves by assuming they are
working on a core native component and always reboot after any
updates associated with the overlayfs adb remount infrastructure.

Test: adb-remount-test.sh
Bug: 127867895
Change-Id: Ic0d37f4754d230c9ac45d6eccb066db277c6d5c6