libcutils: android_get_control_file needs an extra readlink.
If the path to android_get_control_file is a symlink, the final sanity
check will fail that the fd does not have the same path as the given
file. We can't expect callers to readlink() because this would change
the environment key. Instead, try to call readlink on both paths.

Bug: 122556707
Test: "file /dev/block/by-name/userdata" in .rc script
Change-Id: I0df10d7dbe3e572b8335faad812e5cd80bff1733