Update patch set 3
Patch Set 3:

Would it be possible to modify this slightly?

For example, we are thinking about changing the way the compressed memory is stored, so I think it would be useful to have a new Memory object, something like MemoryCompressed. Then you could have two implementations, one that aborts, or always returns an error, and the real one. For places that don't support the compressions stuff, they get the error implementation.

Does this seem feasible?

Patch-set: 3
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>