Update patch set 1
Patch Set 1: Code-Review+1

LGTM assuming you've checked that none of the users of this code are relying on the buffer size to be small. For example, if a user of FrameworkListener reads a command into a buffer that is only 2k long, this change could cause an overflow.

You'll also need to fix NativeMultinetworkJni.cpp because some of the test cases there assert that long DNS queries result in EMSGSIZE.

Patch-set: 1
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1