zip_writer.cc: zero is not a power of 2
Special case the situation where alignment is specified as zero. Buggy
versions of powerof2 would return true for powerof2(0), which is
mathmatically incorrect. To preserve the current behavior, add the
special case.

Fix the version of powerof2() in this code. This version is unused, as
bionic supplies the powerof2 implementation.

This change is necessary to enable integer sanitization on this code.

Somewhat related to Bug: 122975762
Test: atest ziparchive-tests

Change-Id: Id00d48a324e827096e777612d7640dc612492455