init: mount_handler: system-as-root (legacy)
Did not determine dev.mnt.blk.root value on system-as-root, triaged
as /dev/root mount existed before the associated device was
instantiated by ueventd.  As a result when the device shows up the
root mount is updated late when the next mount occurs, which we will
accept, but also the entries are added before removed which causes
the ultimate property entry for root to report empty.

Fix is to change to Remove before Adding.

On system-as-root as well, can not just use fstab.  Determine if a
dm-verity reference is wrapped around system and use that instead.

Add some additional filtration of loop and APEX mounts to reduce
property noise.

Fix issue with creating the std:string line holder from getline(3).

Test: manual on marlin
Bug: 124072565
Change-Id: Ief2e1a6f559cbcbc87273fc2db35c675bb972f43