init: mount_handler: system-as-root
Did not determine dev.mnt.blk.root value on system-as-root, as
discovered because /dev/root mount existed before the device was
instantiated by ueventd.  As a result when the device shows up the
root mount is updated late when the next mount occurs, which we will
accept, but the problem is entries are added before removed which
causes the property entry to report empty due to the change.

Fix is to change to Remove before Adding.

Add some additional filtration of loop and APEX mounts.

Test: manual on marlin
Bug: 124072565
Change-Id: Ief2e1a6f559cbcbc87273fc2db35c675bb972f43