Update patch set 1
Patch Set 1:

This change must be submitted together with the change in PhoneWindowManager. But the change in PhoneWindowManager is hitting a merge conflict with internal master.

Would it be okay if I cherry-picked both changes into internal master, and then submitted this in aosp with a Merged-In? This would cause internal master and aosp to diverge in system/core for a little bit (maybe 1-2 days, due to presubmit).
I don't see any other good way of addressing this, because presubmit will never work until there is a Merged-In, and I cannot merge the frameworks/base change by itself into internal master.

Please advise.

Patch-set: 1
Reviewer: Gerrit User 1149384 <1149384@85c56323-6fa9-3386-8a01-6480fb634889>