<cutils/ashmem.h>: Add ashmem_get_prot_region().
This CL adds a new function to retrieve the current
protection flags for a given shared memory region.

Note that for ashmem regions, this uses an ioctl
that retrieves the region's protection mask directly.

While for memfd regions, the value is synthetized
from the set of seals currently applied to the region.
See technical note for limitations on the result value,
especially regarding PROT_EXEC handling.

BUG=123995193
Change-Id: I3b637447422bfc170198364f83902c77ec240b78