Let blkio cgroup follow cpuset cgroup only
Some app may have different settings in cpuset and schedtune for its own
threads, so let blkio follow cpuset only, which presents the app's state
more accurately.

ex: an app's main thread is in top-app, but some thread of it has
different settings in schedtune and cpuset, and blkio follows schedtune
because it is called later.

Main thread:
6:schedtune:/top-app
5:memory:/
4:cpuset:/top-app
3:cpuacct:/uid_1000/pid_8766
2:cpu:/
1:blkio:/
0::/

Some thread:
6:schedtune:/background
5:memory:/
4:cpuset:/top-app
3:cpuacct:/uid_1000/pid_8766
2:cpu:/
1:blkio:/background
0::/

Bug: 117857342
Test: blkio has same settings with cpuset
Change-Id: I9a140c7d9d93e1dd43c34c8cf066f4a62e2bf604